Refactor component code
